Edit stylesheet of block-and-inline/02-margin-and-padding-2 exercise
Second Assignment of Block and Inline CSS Foundations - Completed
This exercise is real deal among the previous ones. It requires proper and smart use of
margin and padding properties to align elements and give a well-structured layout.
It basic card creation layout with a title, content area, and a button but have all
important concepts for real-world applications.